+++ tests/testapp/test_settextfield.py
@@ -9,6 +9,9 @@ from django.core import serializers
 from django.db import models
 from django.db.migrations.writer import MigrationWriter
 from django.db.models import Q
+from django.db.models import Value
+from django.db.models.functions import Concat
+from django.db.models.functions import Upper
 from django.test import SimpleTestCase
 from django.test import TestCase
 from django.test.utils import isolate_apps
@@ -117,6 +120,19 @@ class TestSaveLoad(TestCase):
 
         not_mouldy = BigCharSetModel.objects.exclude(**{lname: "mouldy"})
         assert not_mouldy.count() == 0
+
+    def test_char_lookup_contains_with_params(self):
+        # Regression test for bug where parameters were in the wrong order
+        instance = BigCharSetModel.objects.create(field={"MOULDY", "rotten"})
+
+        mouldy = BigCharSetModel.objects.annotate(
+            ufield=Upper(
+                Concat("field", Value("")),
+                output_field=SetTextField(base_field=models.CharField(max_length=8)),
+            )
+        ).filter(ufield__contains=Upper(Value("Mouldy")))
+
+        assert set(mouldy) == {instance}
 
     def test_char_len_lookup_empty(self):
         mymodel = BigCharSetModel.objects.create(field=set())